Update README.md and rename components. (#9)
